Get an item from a nlmixr core object

Get an item from a nlmixr core object

Usage

nmObjGet(x, ...)

# S3 method for default
nmObjGet(x, ...)

# S3 method for modelName
nmObjGet(x, ...)

# S3 method for cor
nmObjGet(x, ...)

# S3 method for omegaR
nmObjGet(x, ...)

# S3 method for phiR
nmObjGet(x, ...)

# S3 method for phiSE
nmObjGet(x, ...)

# S3 method for phiRSE
nmObjGet(x, ...)

# S3 method for dataSav
nmObjGet(x, ...)

# S3 method for idLvl
nmObjGet(x, ...)

# S3 method for covLvl
nmObjGet(x, ...)

# S3 method for saemTransformedData
nmObjGet(x, ...)

# S3 method for parHistStacked
nmObjGet(x, ...)

# S3 method for md5
nmObjGet(x, ...)

# S3 method for notes
nmObjGet(x, ...)

# S3 method for sigma
nmObjGet(x, ...)

# S3 method for coefficients
nmObjGet(x, ...)

# S3 method for env
nmObjGet(x, ...)

# S3 method for condition
nmObjGet(x, ...)

# S3 method for simInfo
nmObjGet(x, ...)

# S3 method for seed
nmObjGet(x, ...)

# S3 method for saemCfg
nmObjGet(x, ...)

# S3 method for estimationModel
nmObjGet(x, ...)

# S3 method for atol
nmObjGet(x, ...)

# S3 method for rtol
nmObjGet(x, ...)

# S3 method for maxstepsOde
nmObjGet(x, ...)

# S3 method for hmin
nmObjGet(x, ...)

# S3 method for hmax
nmObjGet(x, ...)

# S3 method for hini
nmObjGet(x, ...)

# S3 method for maxordn
nmObjGet(x, ...)

# S3 method for maxords
nmObjGet(x, ...)

# S3 method for methodOde
nmObjGet(x, ...)

# S3 method for covsInterpolation
nmObjGet(x, ...)

# S3 method for control
nmObjGet(x, ...)

# S3 method for simulationModel
nmObjGet(x, ...)

# S3 method for rxControl
nmObjGet(x, ...)

# S3 method for rxControlWithVar
nmObjGet(x, ...)

# S3 method for saemDopredIpred
nmObjGet(x, ...)

# S3 method for saemDopredPred
nmObjGet(x, ...)

Arguments

x

A specialized list with: - First argument is a nlmixrFitCore environment - Second argument is if the exact argument is requested - The class would be the requested argument name followed by the class "nmObjGet"

...

Other arguments

Value

Value of argument or NULL

Author

Matthew L. Fidler